requirements.txt
openai
python-dotenv
praw
yfinance
beautifulsoup4
requests
serpapi
